skeleton

+++ cu_v4/common/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_flash.c
@@ -1023,7 +1023,7 @@ void FLASH_OB_WRPConfig(uint32_t OB_WRP, FunctionalState NewState)
   { 
     if(NewState != DISABLE)
     {
-      *(__IO uint16_t*)OPTCR_BYTE2_ADDRESS &= (~OB_WRP);
+      *(__IO uint16_t*)OPTCR_BYTE2_ADDRESS &= (uint16_t)(~OB_WRP);
     }
     else
     {
@@ -1065,7 +1065,7 @@ void FLASH_OB_WRP1Config(uint32_t OB_WRP, FunctionalState NewState)
   { 
     if(NewState != DISABLE)
     {
-      *(__IO uint16_t*)OPTCR1_BYTE2_ADDRESS &= (~OB_WRP);
+      *(__IO uint16_t*)OPTCR1_BYTE2_ADDRESS &= (uint16_t)(~OB_WRP);
     }
     else
     {
@@ -1150,7 +1150,7 @@ void FLASH_OB_PCROPConfig(uint32_t OB_PCROP, FunctionalState NewState)
     }
     else
     {
-      *(__IO uint16_t*)OPTCR_BYTE2_ADDRESS &= (~OB_PCROP);
+      *(__IO uint16_t*)OPTCR_BYTE2_ADDRESS &= (uint16_t)(~OB_PCROP);
     }
   }
 }
@@ -1187,7 +1187,7 @@ void FLASH_OB_PCROP1Config(uint32_t OB_PCROP, FunctionalState NewState)
     }
     else
     {
-      *(__IO uint16_t*)OPTCR1_BYTE2_ADDRESS &= (~OB_PCROP);
+      *(__IO uint16_t*)OPTCR1_BYTE2_ADDRESS &= (uint16_t)(~OB_PCROP);
     }
   }
 }
@@ -1284,7 +1284,7 @@ void FLASH_OB_BootConfig(uint8_t OB_BOOT)
   assert_param(IS_OB_BOOT(OB_BOOT));
 
   /* Set Dual Bank Boot */
-  *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS &= (~FLASH_OPTCR_BFB2);
+  *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS &= (uint8_t)(~FLASH_OPTCR_BFB2);
   *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS |= OB_BOOT;
 
 }
@@ -1305,7 +1305,7 @@ void FLASH_OB_BORConfig(uint8_t OB_BOR)
   assert_param(IS_OB_BOR(OB_BOR));
 
   /* Set the BOR Level */
-  *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS &= (~FLASH_OPTCR_BOR_LEV);
+  *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS &= (uint8_t)(~FLASH_OPTCR_BOR_LEV);
   *(__IO uint8_t *)OPTCR_BYTE0_ADDRESS |= OB_BOR;
 
 }
